docs(example): Improve example site design
