+++ example/index.html
@@ -13,15 +13,15 @@
   />
   <link
     rel="shortcut icon"
-    href="./favicon.png"
+    href="./logo.png"
     type="image/x-icon"
   >
   <link
     rel="icon"
-    href="./favicon.png"
+    href="./logo.png"
     type="image/x-icon"
   >
-  <title>React Simple Modal Example</title>
+  <title>React Simple Modal</title>
 </head>
 
 <body>
