+++ example/package.json
@@ -10,7 +10,8 @@
   },
   "dependencies": {
     "components": "^0.1.0",
-    "react-app-polyfill": "^1.0.0"
+    "react-app-polyfill": "^1.0.0",
+    "react-icons": "^3.10.0"
   },
   "alias": {
     "react": "../node_modules/react",
@@ -25,7 +26,7 @@
     "cssnano": "^4.1.10",
     "parcel": "^1.12.3",
     "postcss-cli": "^7.1.0",
-    "tailwindcss": "^1.2.0",
+    "tailwindcss": "^1.3.5",
     "typescript": "^3.4.5"
   }
 }
