+++ example/tailwind.css
@@ -1,3 +1,14 @@
 @tailwind base;
+
+@import url('https://rsms.me/inter/inter.css');
+html {
+  font-family: 'Inter', sans-serif;
+}
+@supports (font-variation-settings: normal) {
+  html {
+    font-family: 'Inter var', sans-serif;
+  }
+}
+
 @tailwind components;
 @tailwind utilities;
